requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
autopage==0.5.2
cfgv==3.4.0
cliff==4.7.0
cmaes==0.10.0
cmd2==2.4.3
colorlog==6.8.2
distlib==0.3.8
exceptiongroup==1.2.1
filelock==3.14.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.5.0
greenlet==3.0.3
grpcio==1.64.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
identify==2.5.36
idna==3.7
igraph==0.11.5
iniconfig==2.0.0
Jinja2==3.1.4
lightning==2.2.5
lightning-utilities==0.11.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
nodeenv==1.9.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
optuna==2.10.1
packaging==24.0
pandas==2.2.2
pbr==6.0.0
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
prettytable==3.10.0
protobuf==5.27.0
Pygments==2.18.0
pyperclip==1.8.2
pytest==8.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.2.5
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
rootutils==1.0.7
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.13.1
six==1.16.0
SQLAlchemy==2.0.30
stevedore==5.2.0
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
texttable==1.7.0
TinyNeuralNetwork @ git+https://github.com/alibaba/TinyNeuralNetwork.git@8c1f2ce00e9584318092956dcbf99f6eb587992c
tomli==2.0.1
torch==2.3.0
torchmetrics==1.4.0.post0
torchvision==0.18.0
tqdm==4.66.4
triton==2.3.0
typing_extensions==4.12.0
tzdata==2024.1
virtualenv==20.26.2
wcwidth==0.2.13
Werkzeug==3.0.3
yarl==1.9.4
